updates to make lib cross platform

+++ Tests/MySQL/Mocks.swift
@@ -15,7 +15,7 @@ public class MockMySQLConnection : MySQLConnectionProtocol {
 
   public var nextResultReturn:CMySQLRow? = nil
 
-  private var uuid: UInt32
+  private var uuid: UInt64
 
   public func equals(otherObject: MySQLConnectionProtocol) -> Bool {
     return uuid == (otherObject as! MockMySQLConnection).uuid
